pyproject.toml
[project]
name = "bigft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chz>=0.4.0",
    "datasets>=4.5.0",
    "peft>=0.18.1",
    "python-dotenv>=1.2.1",
    "tinker>=0.9.0",
    "tinker-cookbook>=0.1.0",
]
